Break prioritized out into its own file
